Rename lookup mixin file, and fix placeholder in flexberry-lookup
tfs #105375
