+++ addon/components/flexberry-lookup.js
@@ -435,7 +435,10 @@ export default FlexberryBaseComponent.extend({
   buildDisplayValue() {
     let selectedModel = this.get('value');
     if (!selectedModel) {
+      this.set('placeholder', t('flexberry-lookup.placeholder'));
       return '';
+    } else {
+      this.set('placeholder', '');
     }
 
     return selectedModel.get(this.get('displayAttributeName'));
